Use useParams for url data
